Make navigation routes more react like
Here we remove instances of href and wrap those components in LinkContainers. This allows our app to utilize client-side rendering and prevent unnecessary reloads making our app faster. We allow the Gallery link to keep  its href since it needs to be in sync with the server. Ideally the Gallery component would self-update and the entire application would not need to be reloaded; however this will be left as a task for later.